requirements.txt
Flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
google-cloud-storage==2.16.0
python-dotenv==1.0.1
google-generativeai==0.7.1
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-mail==0.9.1
bcrypt==4.1.3
pg8000==1.31.2
cloud-sql-python-connector==1.11.0
